// src/types/api.ts
export interface CiserUserInfo {
  id: string;
  displayName: string;
  givenName: string;
  jobTitle: string;
  email: string;
  userPrincipalName: string;
  codigo_bp: string;
  nome_bp: string;
  login_adfs: string;
  is_representante: string;
  erp_email: string;
  data_sincronizacao: string;
  hora_sincronizacao: string;
}